Screen5 completed - can set and save fuel settings OK. Also enhanced pump setters/getters to use float. Added selection box methods to CScreen

+++ Arduino/BTCDieselHeater/Screen1.h
@@ -40,7 +40,7 @@ class CScreen1 : public CScreen
   void showRunState();
   void showThermometer(float desired, float actual);
   void showBodyThermometer(int actual);
-  void showGlowPlug(int power);
+  void showGlowPlug(float power);
   void showFan(int RPM);
   void showFuel(float rate);
   void showRunState(int state, int errstate);
